Make UTF-8 encoding of sitemap.xml explicit
The sitemap specification states that UTF-8 encoding MUST be used. The
library declars this in the XML header, but the spit command depends on
the writer choosing UTF-8 as the encoding. A small change to pass the
encoding through in case the underlying behaviour changes or it doesn’t
work on Windows platforms.

See http://www.sitemaps.org/faq.html#faq_output_encoding