Add validate-sitemap function to core
This commit includes one minor change to the generate-sitemap-and-save
function of version 0.2.2 of the original API. It saved the XML to disk
and returned the XML string. I have changed this to a save-sitemap
function which returns the given File object, which then allows
function chaining to the validator. See the Validation section of the
Readme for an example.